// web/src/api/generated/api.ts
export interface CreatePatientCommand {
    'fullName'?: string | null;
    'age'?: number;
    'gender'?: Gender;
    'note'?: string | null;
}



export const Gender = {
    Male: 'Male',
    Female: 'Female',
    Other: 'Other'
} as const;

export type Gender = typeof Gender[keyof typeof Gender];
// ...
export interface LoginCommand {
    'username'?: string | null;
    'password'?: string | null;
}
// ...
export interface PatientModel {
    'fullName'?: string | null;
    'age'?: number;
    'gender'?: Gender;
    'note'?: string | null;
}


export interface PrescriptionModel {
    'drugId'?: string | null;
    'dosage'?: string | null;
    'quantity'?: number;
    'unitId'?: string | null;
}
export interface RefreshTokenCommand {
    'refreshToken'?: string | null;
}
export interface RefreshTokenResponse {
    'accessToken': string | null;
    'expiresIn'?: number;
    'refreshToken': string | null;
}
export interface RefreshTokenResponseApiResponse {
    'status'?: number;
    'message'?: string | null;
    'results'?: RefreshTokenResponse;
}
export interface Ulid {
    'random'?: string | null;
    'time'?: string;
}
export interface VisitModel {
    'symptoms'?: string | null;
    'diagnosis'?: string | null;
    'note'?: string | null;
    'prescriptionModel'?: Array<PrescriptionModel> | null;
}
// ...
export class ClientApi extends BaseAPI {
    /**
     * 
     * @summary Logging in Account
     * @param {LoginCommand} [loginCommand] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiAuthLoginPost(loginCommand?: LoginCommand, options?: RawAxiosRequestConfig) {
        return ClientApiFp(this.configuration).apiAuthLoginPost(loginCommand, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Logging in Account
     * @param {RefreshTokenCommand} [refreshTokenCommand] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiAuthRefreshTokenPost(refreshTokenCommand?: RefreshTokenCommand, options?: RawAxiosRequestConfig) {
        return ClientApiFp(this.configuration).apiAuthRefreshTokenPost(refreshTokenCommand, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {number} [page] 
     * @param {number} [pageSize] 
     * @param {string} [keyword] 
     * @param {Array<string>} [targets] 
     * @param {string} [sort] 
     * @param {any} [filter] 
     * @param {Array<string>} [originFilters] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiPatientsGet(page?: number, pageSize?: number, keyword?: string, targets?: Array<string>, sort?: string, filter?: any, originFilters?: Array<string>, options?: RawAxiosRequestConfig) {
        return ClientApiFp(this.configuration).apiPatientsGet(page, pageSize, keyword, targets, sort, filter, originFilters, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiPatientsIdGet(id: string, options?: RawAxiosRequestConfig) {
        return ClientApiFp(this.configuration).apiPatientsIdGet(id, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @param {string} id 
     * @param {PatientModel} [patientModel] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiPatientsIdPut(id: string, patientModel?: PatientModel, options?: RawAxiosRequestConfig) {
        return ClientApiFp(this.configuration).apiPatientsIdPut(id, patientModel, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Create a new patient
     * @param {CreatePatientCommand} [createPatientCommand] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiPatientsPost(createPatientCommand?: CreatePatientCommand, options?: RawAxiosRequestConfig) {
        return ClientApiFp(this.configuration).apiPatientsPost(createPatientCommand, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary List visits by patient
     * @param {string} patientId 
     * @param {number} [page] 
     * @param {number} [pageSize] 
     * @param {string} [keyword] 
     * @param {Array<string>} [targets] 
     * @param {string} [sort] 
     * @param {any} [filter] 
     * @param {Array<string>} [originFilters] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiVisitsPatientIdVisitsGet(patientId: string, page?: number, pageSize?: number, keyword?: string, targets?: Array<string>, sort?: string, filter?: any, originFilters?: Array<string>, options?: RawAxiosRequestConfig) {
        return ClientApiFp(this.configuration).apiVisitsPatientIdVisitsGet(patientId, page, pageSize, keyword, targets, sort, filter, originFilters, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Create visit for a patient
     * @param {string} patientId 
     * @param {VisitModel} [visitModel] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public apiVisitsPatientIdVisitsPost(patientId: string, visitModel?: VisitModel, options?: RawAxiosRequestConfig) {
        return ClientApiFp(this.configuration).apiVisitsPatientIdVisitsPost(patientId, visitModel, options).then((request) => request(this.axios, this.basePath));
    }
}